Core: Move is_client from common.client to common.config; fix #7602
This fixes a hidden circular import between common.client and common.config.

This circular import is currently 'hidden' by doing the common.client import within config_get,
but it breaks as soon as common.client attempts to import any other rucio module that depends on config_get (i.e. a large part of the codebase).

common.config is clearly one of the first entrypoints in Rucio, whereas common.client is a secondary module.
To break this circular import, the dependency on common.client is removed by moving is_client to common.config, for 2 reasons:
	1. is_client simply checks os.environ and the config, so despite its name, it's a fairly 'config-heavy' function
	2. as most modules depend on common.config, it's best if common.config depends on as few modules as possible.
